+++ src/main/java/net/servicestack/idea/INativeTypesHandler.java
@@ -18,4 +18,6 @@ public interface INativeTypesHandler {
     String getRelativeTypesUrl();
     NativeTypesLanguage getTypesLanguage();
     String getLanguageUrlName();
+    String getOptionsCommentStart();
+    String getOptionsCommentEnd();
 }