+++ src/main/java/net/servicestack/idea/TypeScriptNativeTypesHandler.java
@@ -18,4 +18,14 @@ public class TypeScriptNativeTypesHandler extends BaseNativeTypesHandler {
     public NativeTypesLanguage getTypesLanguage() {
         return NativeTypesLanguage.TypeScript;
     }
+
+    @Override
+    public String getOptionsCommentStart() {
+        return "*/";
+    }
+
+    @Override
+    public String getOptionsCommentEnd() {
+        return "*/";
+    }
 }